refactor(db): move relay member store ownership
